Add basic pinning test for jsdoc completions
For #82874
